refac: move collapsed member card into separate component
