+++ src/components/cards/MemberCard/index.js
@@ -1,0 +1,3 @@
+import MemberCard from './MemberCard';
+
+export default MemberCard;
